+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [4.2.0] - 2019-04-22
+
+### Added
+
+-   Settings property to use id only with `hashCode()` and `equals()`
+-   Settings property to use String, Primitive and Primitive Wrappers only for `toString()`
+
 ## [4.1.0] - 2019-02-11
 
 ### Added